requirements.txt
pyenchant==1.6.5
